cron job, get rid of route
- get rid of testing html files as well
