+++ server/src/admin/admin-auth.ts
@@ -192,11 +192,6 @@ export const adminRoutes = async (fastify: FastifyInstance, _options: RegisterOp
     },
     required: ["commentId"]
   };
-  const fetchRecentCommentsParams = {
-    type: "object",
-    properties: {},
-    required: []
-  };
 
   fastify.addHook("preHandler", verifyAdmin);
   fastify.addHook("preHandler", verifyRecaptcha);
