--- server/src/template.html
@@ -1,194 +1,0 @@
-<!DOCTYPE html>
-  <html>
-  <head>
-  
-    <meta charset="utf-8">
-    <meta http-equiv="x-ua-compatible" content="ie=edge">
-    <title>Comment Update</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <style type="text/css">
-    /**
-     * Google webfonts. Recommended to include the .woff version for cross-client compatibility.
-     */
-    @media screen {
-      @font-face {
-        font-family: 'Source Sans Pro';
-        font-style: normal;
-        font-weight: 400;
-        src: local('Source Sans Pro Regular'), local('SourceSansPro-Regular'), url(https://fonts.gstatic.com/s/sourcesanspro/v10/ODelI1aHBYDBqgeIAH2zlBM0YzuT7MdOe03otPbuUS0.woff) format('woff');
-      }
-      @font-face {
-        font-family: 'Source Sans Pro';
-        font-style: normal;
-        font-weight: 700;
-        src: local('Source Sans Pro Bold'), local('SourceSansPro-Bold'), url(https://fonts.gstatic.com/s/sourcesanspro/v10/toadOcfmlt9b38dHJxOBGFkQc6VGVFSmCnC_l7QZG60.woff) format('woff');
-      }
-    }
-    /**
-     * Avoid browser level font resizing.
-     * 1. Windows Mobile
-     * 2. iOS / OSX
-     */
-    body,
-    table,
-    td,
-    a {
-      -ms-text-size-adjust: 100%; /* 1 */
-      -webkit-text-size-adjust: 100%; /* 2 */
-    }
-    /**
-     * Remove extra space added to tables and cells in Outlook.
-     */
-    table,
-    td {
-      mso-table-rspace: 0pt;
-      mso-table-lspace: 0pt;
-    }
-    /**
-     * Better fluid images in Internet Explorer.
-     */
-    img {
-      -ms-interpolation-mode: bicubic;
-    }
-    /**
-     * Remove blue links for iOS devices.
-     */
-    a[x-apple-data-detectors] {
-      font-family: inherit !important;
-      font-size: inherit !important;
-      font-weight: inherit !important;
-      line-height: inherit !important;
-      color: inherit !important;
-      text-decoration: none !important;
-    }
-    /**
-     * Fix centering issues in Android 4.4.
-     */
-    div[style*="margin: 16px 0;"] {
-      margin: 0 !important;
-    }
-    body {
-      width: 100% !important;
-      height: 100% !important;
-      padding: 0 !important;
-      margin: 0 !important;
-    }
-    /**
-     * Collapse table borders to avoid space between cells.
-     */
-    table {
-      border-collapse: collapse !important;
-    }
-    a {
-      color: #1a82e2;
-    }
-    img {
-      height: auto;
-      line-height: 100%;
-      text-decoration: none;
-      border: 0;
-      outline: none;
-    }
-    </style>
-  
-  </head>
-  <body style="background-color: #e9ecef;">
-  
-    <!-- start preheader -->
-    <div class="preheader" style="display: none; max-width: 0; max-height: 0; overflow: hidden; font-size: 1px; line-height: 1px; color: #fff; opacity: 0;">
-      ${preHeader}
-    </div>
-    <!-- end preheader -->
-  
-    <!-- start body -->
-    <table border="0" cellpadding="0" cellspacing="0" width="100%">
-  
-      <!-- start logo -->
-      <tr>
-        <td align="center" bgcolor="#e9ecef">
-          <table border="0" cellpadding="0" cellspacing="0" width="100%" style="max-width: 600px;">
-            <tr>
-              <td align="center" valign="top" style="padding: 36px 24px;">
-                <a href="https://dev.timescalecreator.org" target="_blank" style="display: inline-block;">
-                  <img src="https://dev.timescalecreator.org/assets/TSCreatorLogo-BBvqDlKG.png" alt="Logo" border="0" width="48" style="display: block; width: 48px; max-width: 48px; min-width: 48px;">
-                </a>
-              </td>
-            </tr>
-          </table>
-        </td>
-      </tr>
-      <!-- end logo -->
-  
-      <!-- start hero -->
-      <tr>
-        <td align="center" bgcolor="#e9ecef">
-          <table border="0" cellpadding="0" cellspacing="0" width="100%" style="max-width: 600px;">
-            <tr>
-              <td align="left" bgcolor="#ffffff" style="padding: 36px 24px 0; font-family: 'Source Sans Pro', Helvetica, Arial, sans-serif; border-top: 3px solid #d4dadf;">
-                <h1 style="margin: 0; font-size: 28px; font-weight: 600; letter-spacing: -1px; line-height: 48px;">${title}</h1>
-              </td>
-            </tr>
-          </table>
-        </td>
-      </tr>
-      <!-- end hero -->
-  
-      <!-- start comment block -->
-      <tr>
-        <td align="center" bgcolor="#e9ecef">
-          <table border="0" cellpadding="0" cellspacing="0" width="100%" style="max-width: 600px;">
-  
-           
-            <tr>
-              <td align="left" bgcolor="#ffffff" style="padding: 0 24px 16px 24px; font-family: 'Source Sans Pro', Helvetica, Arial, sans-serif; font-size: 16px; line-height: 24px;">
-                <p style="margin: 0;">${message}</p>
-              </td>
-            </tr>
-             <!-- start comment -->
-            ${commentsHtml}
-            <!-- end comment -->
-              <!-- start button -->
-              <tr>
-                <td align="left" bgcolor="#ffffff">
-                  <table border="0" cellpadding="0" cellspacing="0" width="100%">
-                    <tr>
-                      <td align="center" bgcolor="#ffffff" style="padding: 12px;">
-                        <table border="0" cellpadding="0" cellspacing="0">
-                          <tr>
-                            <td align="center" bgcolor="#1a82e2" style="border-radius: 6px;">
-                              <a href="${link}" target="_blank" style="display: inline-block; padding: 16px 36px; font-family: 'Source Sans Pro', Helvetica, Arial, sans-serif; font-size: 16px; color: #ffffff; text-decoration: none; border-radius: 6px;">${buttonText}</a>
-                            </td>
-                          </tr>
-                        </table>
-                      </td>
-                    </tr>
-                  </table>
-                </td>
-              </tr>
-              <!-- end button -->
-              <!-- start link text -->
-              <tr>
-                <td align="left" bgcolor="#ffffff" style="padding: 24px; font-family: 'Source Sans Pro', Helvetica, Arial, sans-serif; font-size: 16px; line-height: 24px;">
-                  <p style="margin: 0;">If that doesn't work, copy and paste the following link in your browser:</p>
-                  <p style="margin: 0;"><a href="${link}" target="_blank">${link}</a></p>
-                </td>
-              </tr>
-              <!-- end link text -->
-            
-          </table>
-        </td>
-      </tr>
-      <!-- end comment block -->
-  
-      <!-- start footer -->
-      <tr>
-        <td align="center" bgcolor="#e9ecef" style="padding: 24px;">
-        </td>
-      </tr>
-      <!-- end footer -->
-  
-    </table>
-    <!-- end body -->
-  
-  </body>
-</html>
